feat: relasi dan seeder
Menambah relasi pada model dan seeder untuk data dummy di database

+++ database/migrations/2024_09_11_023135_create_accounts_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->string('account_number')->unique();
-            $table->decimal('balance', 15, 2);
+            $table->bigInteger('balance');
             $table->timestamps();
         });
     }
